add `make_random_graph`
